feat(help): Show help for targets in all Makefiles
Rather than only documenting the targets for `micromamba.mk`.  This
makes it so that the user can include custom rules in their own
Makefiles and still benefit from `make help`, e.g.:

    # file: Makefile
    include micromamba.mk
    include user_makefile.mk

    # file: user_makefile.mk
    foo: bar ## some docstring
    	echo something

When running `make help` it will not only show the help messages for the
targets in `micromamba.mk` but also the user's custom makefile
`user_makefile.mk` (and `Makefile` itself), provided that the targets
are documented with `target: ... ## <docstring>`.