feat(cov): Require full line coverage for test modules
Do a second run of `coverage report`, but this time limiting the
coverage calculation to files under the `tests/` package, and overriding
the `--fail-under` argument to 100.

In most cases, dead code in test modules is undesirable.

* In the intentional scenario always skipped or conditionally skipped
  tests, these should be marked with `# pragma: no cover` in the
  scenario which they don't need to be computed (e.g Windows specific,
  or python-version specific tests).
* In the *un*intentional scenario, e.g. assertions inside a loop that is
  unexpectedly iterating over an empty container, this is probably an
  oversight in the test logic.

By failing on uncovered lines, the developer can be alerted and decide
on each case, marking `no-cover` on the unimportant lines or fixing the
unexpected dead statements.

Note:
* The flag `--include 'tests/*'` will include modules in sub-packages as
  well.
* The coverage percentage is calculated based only on the included
  files.
* Running `coverage report` twice is not ideal, but this is much faster
  than running the entire suite of tests, so it's not a major problem.